fixing test for nodes.
In the previous case, 4 is rendered as a child of 1. Except that 4 has a
width of 2 and 1 has a width of 1. The child shouldn't be wider than the
parent. The correct case is that there should be two nodes for 4 of
width 1 each. The first node is a child of 1 and the second one is a
child of 5.